+++ src/Controller/Ajax/DynamicCustomFieldController.php
@@ -17,9 +17,12 @@ use Auth;
 use Eventum\CustomField\Fields\DynamicCustomFieldInterface;
 use Eventum\Db\Doctrine;
 use Eventum\Model\Repository\CustomFieldRepository;
+use LazyProperty\LazyPropertiesTrait;
 
 class DynamicCustomFieldController extends AjaxBaseController
 {
+    use LazyPropertiesTrait;
+
     /** @var string */
     protected $tpl_name = 'js/dynamic_custom_field.tpl.js';
     /** @var int */
@@ -29,18 +32,21 @@ class DynamicCustomFieldController extends AjaxBaseController
     /** @var string */
     private $form_type;
     /** @var CustomFieldRepository */
-    private $repo;
+    private $customFieldRepository;
 
     /**
      * {@inheritdoc}
      */
     protected function configure(): void
     {
+        $this->initLazyProperties([
+            /** @see getCustomFieldRepository */
+            'customFieldRepository',
+        ]);
         $request = $this->getRequest();
 
         $this->issue_id = $request->request->getInt('iss_id') ?: $request->query->getInt('iss_id');
         $this->form_type = $request->request->get('form_type') ?: $request->query->get('form_type');
-        $this->repo = Doctrine::getCustomFieldRepository();
     }
 
     /**
@@ -63,9 +69,9 @@ class DynamicCustomFieldController extends AjaxBaseController
     private function getData(): array
     {
         if ($this->issue_id) {
-            $customFields = $this->repo->getListByIssue($this->prj_id, $this->issue_id, $this->role_id, null, true);
+            $customFields = $this->customFieldRepository->getListByIssue($this->prj_id, $this->issue_id, $this->role_id, null, true);
         } else {
-            $customFields = $this->repo->getListByProject($this->prj_id, $this->role_id, $this->form_type, null, true);
+            $customFields = $this->customFieldRepository->getListByProject($this->prj_id, $this->role_id, $this->form_type, null, true);
         }
 
         $data = [];
@@ -94,5 +100,10 @@ class DynamicCustomFieldController extends AjaxBaseController
     protected function prepareTemplate(): void
     {
         $this->tpl->assign('fields', $this->getData());
+    }
+
+    private function getCustomFieldRepository(): CustomFieldRepository
+    {
+        return Doctrine::getCustomFieldRepository();
     }
 }