+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 See [Upgrading] for details on how to upgrade.
 
+- Fix `DynamicCustomFieldController` accessing repository too soon, #1041
+
 [3.10.3]: https://github.com/eventum/eventum/compare/v3.10.2...master
 
 ## [3.10.2] - 2021-04-01